+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Fetch Youtube Transcript Changelog
 
-## [0.1.1] - 2024-11-26
+## [1.1.2] - 2024-01-02
+
+### Changed
+
+- Improved transcript filename generation to use video title instead of video ID
+- Added filename sanitization to handle special characters in video titles
+
+## [1.1.1] - 2024-11-26
 
 ### Added